+++ components/bootstrappers/neo4j/Cargo.toml
@@ -23,16 +23,26 @@ repository = "https://github.com/drasi-project/drasi-core"
 keywords = ["drasi", "plugin", "bootstrap", "neo4j"]
 categories = ["development-tools"]
 
+[lib]
+crate-type = ["lib", "cdylib"]
+
 [lints]
 workspace = true
 
 [dependencies]
 drasi-lib.workspace = true
 drasi-core.workspace = true
+drasi-plugin-sdk = { workspace = true }
+utoipa = { workspace = true }
 anyhow = "1.0"
 async-trait = "0.1"
 chrono = "0.4"
 log = "0.4"
 ordered-float = "3.0"
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.0", features = ["full"] }
 neo4rs = "0.8"
+
+[features]
+dynamic-plugin = []
