+++ components/bootstrappers/neo4j/src/config.rs
@@ -1,0 +1,26 @@
+// Copyright 2026 The Drasi Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Configuration types for the Neo4j bootstrap provider.
+
+/// Configuration for the Neo4j bootstrap provider.
+#[derive(Debug, Clone, PartialEq)]
+pub struct Neo4jBootstrapConfig {
+    pub uri: String,
+    pub user: String,
+    pub password: String,
+    pub database: String,
+    pub labels: Vec<String>,
+    pub rel_types: Vec<String>,
+}
