+++ components/sources/neo4j/Cargo.toml
@@ -23,12 +23,17 @@ repository = "https://github.com/drasi-project/drasi-core"
 keywords = ["drasi", "plugin", "neo4j", "cdc"]
 categories = ["database"]
 
+[lib]
+crate-type = ["lib", "cdylib"]
+
 [lints]
 workspace = true
 
 [dependencies]
 drasi-lib.workspace = true
 drasi-core.workspace = true
+drasi-plugin-sdk = { workspace = true }
+utoipa = { workspace = true }
 anyhow = "1.0"
 async-trait = "0.1"
 chrono = "0.4"
@@ -48,3 +53,6 @@ serial_test = "3.0"
 testcontainers = "0.23"
 tokio-test = "0.4"
 uuid = { version = "1.0", features = ["v4"] }
+
+[features]
+dynamic-plugin = []
